requirements.txt
amqp==5.1.1
ansicon==1.89.0
APScheduler==3.9.1
arrow==1.3.0
asgiref==3.5.2
beem-africa==0.1.2
billiard==3.6.4.0
blessed==1.20.0
celery==5.2.7
certifi==2022.6.15.2
charset-normalizer==2.1.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
distlib==0.3.7
Django==4.1.1
django-admin-interface==0.20.0
django-colorfield==0.7.2
django-crontab==0.7.1
django-etc==1.3.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-grappelli==3.0.3
django-hitcount==1.3.5
django-mathfilters==1.0.0
django-picklefield==3.1
django-q==1.3.9
django-social-share==2.3.0
django-tinymce==3.5.0
filelock==3.12.4
idna==3.4
jinxed==1.2.0
kombu==5.3.2
Pillow==9.2.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
PyJWT==2.5.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2023.3.post1
redis==3.5.3
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
types-python-dateutil==2.8.19.14
tzdata==2023.3
tzlocal==5.1
urllib3==1.26.12
vine==5.0.0
virtualenv==20.24.5
wcwidth==0.2.8
whitenoise==6.2.0
